requirements.txt
psycopg2
python-dotenv
djangorestframework
Pillow
django-filter
djangorestframework-simplejwt 
requests
django-cors-headers
drf-yasg
redis
